requirements.txt
Django>=3.0,<4.0
djangorestframework
python-arango
django-cors-headers
djangorestframework-simplejwt